requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
phonenumbers>=8.13.0
python-dateutil>=2.8.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0

